refactor: Move node fns, entry fn codegen into codegen module
